requirements.txt.txt
opencv-python==4.8.0.74
streamlit==1.25.0
numpy==1.25.1
pillow==9.5.0